+++ Tests/StreamChatUITests/SnapshotTests/MessageActionsPopup/ChatMessageActionsVC_Tests.swift
@@ -18,7 +18,7 @@ import XCTest
         chatMessageController = .mock()
         vc = ChatMessageActionsVC()
         vc.messageController = chatMessageController
-        vc.channel = .mock(cid: .unique, config: .mock(), ownCapabilities: [.sendReply, .quoteMessage, .readEvents])
+        vc.channel = .mock(cid: .unique, config: .mock(), ownCapabilities: [.sendReply, .quoteMessage, .readEvents, .flagMessage])
 
         chatMessageController.simulateInitial(
             message: ChatMessage.mock(id: .unique, cid: .unique, text: "test", author: ChatUser.mock(id: .unique)),
@@ -62,7 +62,7 @@ import XCTest
 
         let vc = TestView()
         vc.messageController = chatMessageController
-        vc.channel = .mock(cid: .unique, config: .mock(), ownCapabilities: [.sendReply, .quoteMessage, .readEvents])
+        vc.channel = .mock(cid: .unique, config: .mock(), ownCapabilities: [.sendReply, .quoteMessage, .readEvents, .flagMessage])
         AssertSnapshot(vc.embedded())
     }
 
@@ -336,9 +336,21 @@ import XCTest
             state: .remoteDataFetched
         )
 
-        vc.channel = .mock(cid: .unique, ownCapabilities: [])
+        vc.channel = .mock(cid: .unique, ownCapabilities: [.flagMessage])
 
         XCTAssertTrue(vc.messageActions.contains(where: { $0 is FlagActionItem }))
+    }
+
+    func test_messageActions_whenMessageIsSentByAnotherUser_whenNoCapability_doesNotContainFlagAction() {
+        chatMessageController.simulateInitial(
+            message: ChatMessage.mock(isSentByCurrentUser: false),
+            replies: [],
+            state: .remoteDataFetched
+        )
+
+        vc.channel = .mock(cid: .unique, ownCapabilities: [])
+
+        XCTAssertFalse(vc.messageActions.contains(where: { $0 is FlagActionItem }))
     }
 
     func test_messageActions_whenSendingFailed_thenContainsResendActionEditActionDeleteAction() {
@@ -419,7 +431,8 @@ import XCTest
                 .sendReply,
                 .readEvents,
                 .updateAnyMessage,
-                .deleteAnyMessage
+                .deleteAnyMessage,
+                .flagMessage
             ]
         )