+++ Tests/StreamChatUITests/SnapshotTests/MessageActionsPopup/ChatMessagePopupVC_Tests.swift
@@ -72,6 +72,10 @@ extension ChatMessagePopupVC: AppearanceProvider {}
         vc.reactionsController = reactionsController
         actionsController = ChatMessageActionsVC()
         actionsController.messageController = chatMessageController
+        actionsController.channel = .mock(
+            cid: .unique,
+            ownCapabilities: [.flagMessage, .sendReply, .quoteMessage, .readEvents]
+        )
         actionsController.channelConfig = .mock()
         vc.actionsController = actionsController
     }